requirements.txt
langchain
langchain-core
langchain-openai
langchain-anthropic
langchain-google-genai
langchain-huggingface
numpy
scikit-learn
transformers
huggingface_hub
python-dotenv
TensorFlow
grandalf
langchain-community
pypdf
beautifulsoup4
langchain-experimental
chromadb
langchain-chroma
tiktoken
openai
langchain-pinecone
pinecone
#pip install -r requirements.txt  